requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
faiss-cpu
pypdf
python-dotenv
slowapi
